+++ pytorch_lightning/strategies/ddp.py
@@ -102,7 +102,7 @@ class DDPStrategy(ParallelStrategy):
             checkpoint_io=checkpoint_io,
             precision_plugin=precision_plugin,
         )
-        log.verbose(f"Initializing DDP: {self.__class__.__name__}")
+        log.detail(f"{self.__class__.__name__}: initializing DDP plugin")
         self.interactive_ddp_procs = []
         self._num_nodes = 1
         self.sync_batchnorm = False
@@ -172,7 +172,9 @@ class DDPStrategy(ParallelStrategy):
 
     def _setup_model(self, model: Module) -> DistributedDataParallel:
         """Wraps the model into a :class:`~torch.nn.parallel.distributed.DistributedDataParallel` module."""
-        return DistributedDataParallel(module=model, device_ids=self.determine_ddp_device_ids(), **self._ddp_kwargs)
+        device_ids = self.determine_ddp_device_ids()
+        log.detail(f"setting up DDP model with device ids: {device_ids}, kwargs: {self._ddp_kwargs}")
+        return DistributedDataParallel(module=model, device_ids=device_ids, **self._ddp_kwargs)
 
     def _call_children_scripts(self):
         # bookkeeping of spawned processes
@@ -244,7 +246,7 @@ class DDPStrategy(ParallelStrategy):
         self._rank_0_has_called_call_children_scripts = True
 
     def setup_distributed(self):
-        log.verbose(f"{self.__class__.__name__}: setting up distributed...")
+        log.detail(f"{self.__class__.__name__}: setting up distributed...")
         reset_seed()
 
         # determine which process we are and world size
@@ -290,6 +292,7 @@ class DDPStrategy(ParallelStrategy):
             self._ddp_kwargs["find_unused_parameters"] = True
 
     def _register_ddp_hooks(self) -> None:
+        log.detail(f"registering ddp hooks")
         # In 1.8, DDP communication hooks only work with NCCL backend and SPSD (single process single device) mode
         # Since 1.9, DDP communication hooks can work on all backends.
         if _TORCH_GREATER_EQUAL_1_9 or (
@@ -309,6 +312,7 @@ class DDPStrategy(ParallelStrategy):
                     self._reinit_optimizers_with_post_localSGD(self._ddp_comm_state.start_localSGD_iter)
 
     def _reinit_optimizers_with_post_localSGD(self, warmup_steps: int):
+        log.detail("reinitializing optimizers with post localSGD")
         optimizers = self.lightning_module.trainer.optimizers
         if self._model_averaging_period is None:
             raise ValueError(
@@ -352,7 +356,7 @@ class DDPStrategy(ParallelStrategy):
         _convert_to_lightning_optimizers(trainer)
 
     def configure_ddp(self) -> None:
-        log.verbose(f"{self.__class__.__name__}: configuring DDP...")
+        log.detail(f"{self.__class__.__name__}: configuring DistributedDataParallel")
         self.pre_configure_ddp()
         self.model = self._setup_model(LightningDistributedModule(self.model))
         self._register_ddp_hooks()
@@ -383,7 +387,7 @@ class DDPStrategy(ParallelStrategy):
             prepare_for_backward(self.model, closure_loss)
 
     def model_to_device(self):
-        log.verbose(f"{self.__class__.__name__}: moving model to device [{self.root_device}]...")
+        log.detail(f"{self.__class__.__qualname__}: moving model to device [{self.root_device}]...")
         self.model.to(self.root_device)
 
     def reduce(self, tensor, group: Optional[Any] = None, reduce_op: Union[ReduceOp, str] = "mean") -> torch.Tensor:
@@ -504,7 +508,7 @@ class DDPStrategy(ParallelStrategy):
         raise DeadlockDetectedException(f"DeadLock detected from rank: {self.global_rank} \n {trace}")
 
     def teardown(self) -> None:
-        log.verbose(f"{self.__class__.__name__}: tearing down plugin...")
+        log.detail(f"{self.__class__.__qualname__}: tearing down DDP plugin")
         super().teardown()
         if isinstance(self.model, DistributedDataParallel):
             self.model = self.lightning_module
@@ -514,7 +518,7 @@ class DDPStrategy(ParallelStrategy):
 
         if self.on_gpu:
             # GPU teardown
-            log.verbose(f"{self.__class__.__name__}: moving model to CPU...")
+            log.detail(f"{self.__class__.__qualname__}: moving model to CPU")
             self.lightning_module.cpu()
             # clean up memory
             torch.cuda.empty_cache()